Major Update
Lots of different updates since it has been a while from the last one. I have templated most of the state space code and it should be much easier to use and adapt to changing inputs and states. Also included the optimizer class to actually optimize the process post model validation.