FIX DAY 0 MISSING VALUES ISSUE
In test_config.yaml,  we added a new Initial field under each state variable. As a result, when a Bioreactor instance is initialized, it contains nominal Day 0 values for all the state variables and will be replaced by the input vector in ingest_vector(). In Bioreactor>ingest_vector(), we modified Line 293 such that only non-NA values in the vector will replace the corresponding variables in Bioreactor.data such that missing values from the input vector will not create missing values in Bioreactor.data. In Bioreactor>get_result(), we removed DailyFeed and TotalFeed and added CurrentTime, FeedRate, and FeedRates to indicate the current time/batch age (day), the optimized nutrient feed rate in mL/min for Kepware, and all feed rates in mL/min from Day 0 to the end of run. We renamed Time to Times to indicate that it's an array of times.