Short Commit 2025-08-29
[MPC Code Integration]
generate_model.py - minor change to display model results
mpc_optimizer.py - Implemented fix to convert manipulated inputs in the vector that are supposed to be normalized to normalized values by dividing by the volume. Need to implement a further fix to ensure the volume is correct when passed in a vector.
visualize.py - minor modification to add constraint lines to the legend

[Data]
B7H4_model_parameters.json - minor update to hidden state variable
Test-Exp.yaml - commented out the end of run variables section since it was throwing error, did not remove in case we needed to see the structure of this for the future. Also commented at some of the bioreactors to make debugging quicker.