Retrained state space model due to instability in matrix. Added weight to prevent optimizer from giving solution with matrix instability determined by eigenvalues > 0. Updated train_model, generate_model, and generate_report based on new weight assigned n json model file. Generated a new report based on this model.
Added weights for overshoot, undershoot, and trajectory discount in the controller section of the yaml file. This can be tuned depending on if the model if prone to overshoot or undershooting the target trajectory. The trajectory discount will weight either the closest or furthest predictions based on its value, useful if you need to ensure the next prediction is highly accurate.