mpc_optimizer: modified ingest vectors to ensure it changes the curr_time variable and outputs the right data. The curr_time was not being changed in the SAM automation test so this is to fix that issue by counting the number of vectors being ingested and setting the curr_time to 1 minus that.
train_model: added print outputs for the stability penalties to see how the evolve during training. Reduced the two gramian penalties to just one for easier use. Also modified how they are calculated, they should be more accurate now.

[Data Files]
Created a new masterdatatable for SAM-1000L experiment that contains CTM1 data scaled down to the 1000L scale so that when testing the controller and model we can directly simulate the typical conditions at the 1K scale.
Created a json file to mimic the data input coming from kafka, it has the exact same naming conventions as well to test the column mapping is working properly. This is in an attempt to fix any issues before they may arise, and especially while the water run is occuring this week.
Retrained the B7H4 model a bit to try to increase stability. Still need to test the model to determine if it is more stable.