[Major Update]
estimation_loop_tech_sim.py - Created an estimation loop file to simulate the exact format and names of data coming from kafka to prepare for the MPC run in SAM lab. This script requires a json file with the correct names for each dictionary. See an example in the SAM-1000L-B7H4-experiment.

functions.py - changed a parameter so it by defaults exports csv files to a "data" folder. You can manually specify the name you want as well if desired.

mpc_optimizer.py - Fixed issues with converting daily feeds. now when ingesting a vector it will not normalize the vector until after all vectors have been ingested. It will then calculate daily feed from the totalizer values unnormalized and then post-calculation it will normalize the values based on the volume. This is because a problem arose where it would normalize the values then calculate the daily from the normalized values but this is incorrect and caused negative feeds in some cases. Also ensured that it appends a 0 post difference calculation so that the feed is on the correct day and not a day ahead, as was the case previously.

[Other updates]
Renamed a few experiments in the data folder for naming consistency which is the reason this commit is so large.