[Crucial Commit]
[data]
Continued testing of new models on CTM data. Created separate CTM test experiements for ease of use. Finalized the B7H4 SAM batch experiment yaml file with the latest model and tuning parameters for the estimator. This will be the model used on the November 21st batch.

[models]
Retrained B7H4 model with much weights on all parameters to achieve model with more accuracy and stability overall
First training of CD19/CD20 model with platform fit ambr DoE data. Achieved very good accuracy for titer and overall summed squares. Generated report to showcase the training and testing fit to data.

[src]
ssm - modified lsim function to include np.clip on the xout values so any negative values are clipped to 0 when predicting the next days state. This helps avoid negative values which are not realistic in the outputs.
train_model - changed the SVD minimum target in the controllability matrix penalty to slightly lower for training the CD19/CD20 model.
generate_model - moved the string variable to determine whether to train the model or output graphs to the top so its easier to access.
generate_report - added instability weights to model training object since this script had not been used in a while.