[Update - First Day of MPC SAM Batch 2025]
[data]
SAM-1000L-B7H4-experiment: Today (11/21/2025) is the first day of the MPC batch. Code was run to generate plots for confirmation that automation is working correctly. All setpoints were updated in kepware and batch is off to a smooth start.

All other experiment commits were for further testing before the actual batch began.

[src]
estimation_loop_tech_sim.py - updating day to 0 for the start of the MPC batch.