[MPC Day 3 and Bug Fix]
[data]
SAM-1000L-B7H4-experiment: day 3 of the mpc SAM experiment. MPC was run manually to validate that setpoints are accurate. Graphs and csv files were generated

[src]
mpc_optimizer.py - issue found on day 2 where the process inputs need to be shifted back one day due to ingest_vectors overwriting the optimizated setpoints. This was fix after the ingest_vectors for loop. The automation code using kafka is not able to update with this fix since it would require redployment.
To revert back to previous code for validation purposes, comment out lines 409 to 415 in this file. All figures in the experiment folder will be based on old code running in automation and not after the bug fix. The bug fix code will be compared to the old code to assess criticality of bug.