[MPC Day 4 & 5]
[data]
Ran optimization for day 4 & 5. There is a larger issue that has become known around error propogation between manually running the code and the automation that runs. The exact reason has not been pinpointed. What is known so far is that the code accumulates optimization errors as the days progress so the optimal values are different between machines. Different python and package versions are being used between machines but this was ruled out as the issue by running it locally with the exact versions the automation code is running.
The secondary issue is the lag between days that was noticed early on but cannot be fixed. This bug does not seem to be as impacful as originally thought.

[src]
changed day to day 4 and 5 in estimation loop.